requirements.txt
tensorflow==2.7.0
numpy==1.19.5
pandas==1.1.5
matplotlib==3.2.2
torch==1.10.0
torchvision==0.11.1
scikit-learn==1.0.2
sewar==0.4.4
opencv-python==4.1.2
Pillow==7.1.2